Paid app: Toolbox ipa file paid for free
I bought this app and then had a jailbroken devices and made it into a ipa file with filza file manager for iOS 12-13. This is a paid app which I bought to share with cydia for people that like Minecraft, this is a mod/toolbox app turned into a ipa file for Minecrafters.