Code Editor Ipa file Paid
I bought code editor, just to share for people that like coding and are developers and also I had to buy this app to turn it into a ipa file with my jailbroken ipad 7 gen iOS 13. Hope you guys like this free app I bought for a expensive price that cost 25 dollars. This app is for coders or people that are staring oof for developing apps. This ipa file works for iOS 12-13.