Worlds.ipa for iOS 11,12,13
This ipa is not one I bought , it is a free ipa file for Minecraft fans or Minecrafters. This is a app just converted into a ipa file for anyone that wants to download it. It works for non jailbroken devices and jailbroken devices running iOS 11,12,13 on iOS. Android users can’t download this unless they have a Apple product.