Attractive Places IPA
This is a app that use to be on the Apple AppStore and it is called Attractive Places. This App is basically hiding the fact that movieboxpro is inside of the app. To make movieboxpro replaced the whole app you guys would have to download this app with altstore if your not Jailbroken and filza file manager if you guys are Jailbroken ,but you guys need to open the app and then once open tap on the + button on the upper right corner and then log in with google drive but make sure you guys logged in on the movieboxpro.app site.