NineAnimator IPA
NineAnimator is a alternative to anime hub or funanimation. Watch free anime, this app is not on the AppStore on iOS devices, supports all iOS version from iOS 11-14. 

Note: Use Altstore or non-Jailbroken to install app or use filza file manager to install the ipa file right to your iOS device if your Jailbroken.