TutuApp
This is TutuApp vip for free patched by me and it works fully and all you guys need is altstore or filza if your Jailbroken and it works fully. Whoever loves TutuApp would love this third party AppStore. Works with all devices that are Jailbroken only to use the apps on TutuApp!

Note: To install use altstore for non-Jailbroken devices! Use filza file manager for Jailbroken devices?